requirements.txt
PySide6>=6.6.0
matplotlib>=3.8.0
psycopg2-binary>=2.9.9
google-generativeai>=0.3.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
